# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth",
        "description": "Document about Authentification ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/episodes/create": {
            "post": {
                "tags": [
                    "create episodes"
                ],
                "summary": "Add episodes",
                "operationId": "1d9365d31563dcb871316524caa8d065",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "audio"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "dev"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "text tets"
                                    },
                                    "audio": {
                                        "type": "string",
                                        "example": "https//images/music.mp3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "episode created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episodes created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "episode not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episode not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/episodes/{id}/update/": {
            "put": {
                "tags": [
                    "update episodes"
                ],
                "summary": "update episodes",
                "operationId": "bdca1ee5e169336a7495d0033668c954",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the episode to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "audio"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Epis"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "text tets"
                                    },
                                    "audio": {
                                        "type": "string",
                                        "example": "https//audio.mp3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "episode update successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episodes update successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "episode not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episode not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/episodes/{id}/delete": {
            "delete": {
                "tags": [
                    "delete episode"
                ],
                "summary": "delete episode",
                "operationId": "f19c3a34be6c4ad8ad4f0fce9dd5d94c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the episode to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "episode delete successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episodes delete successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "episode not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "episodes not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}